Removed Session stuff incl. integrated cookie handling
Removed the Session stuff from the Authenticator trait, because it was not useful for a general approach. For the same reason the automatic setting of a private session cookie was removed. This are two things that are dependent on the usecase of the login and as such should be implemented individually. One way of doing this, is shown in the example.